getMain
