+++ frontend/api/axios.js
@@ -1,7 +1,8 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import axios from "axios";
 
-const baseURL = "http://172.30.1.3:8080";
+const baseURL = "http://10.0.2.16:8080"
+//"http://172.30.1.3:8080";
 // http://현재 wifi의 ip 주소:8080
 
 export const Axios = axios.create({
