+++ frontend/api/getMyPage.js
@@ -4,7 +4,6 @@ export default getMyPage = async (userId, handleSubscribeInfo) => {
 	try {
 		const res = await AxiosToken.get(`/mypage/${userId}`);
 		const data = res.data.data;
-
 		handleSubscribeInfo(data.isSubscribed, data.subscribedInfo, {
 			joinedDate: data.joinedDate,
 			username: data.username,
