+++ frontend/src/pages/MainPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { TouchableOpacity, Image, Alert } from "react-native";
 
 import CommunityHeader from "../organisms/Header";
@@ -78,30 +78,19 @@ const ListTitle = ({ typo }) => {
 	);
 };
 
-const books = [...Array(10).keys()].map((id) => {
-	return {
-		id: id + 1,
-		name: `운영체제${id + 1}`,
-		image: "https://image.yes24.com/goods/89496122/XL",
-	};
-});//api적용시 삭제
 
-const notes = [...Array(10).keys()].map((id) => {
-	return {
-		id: id + 1,
-		name: `필기자료${id + 1}`,
-		image:
-			"https://simage.mujikorea.net/goods/31/11/79/07/4550002435097_N_N_400.jpg",
-	};
-});//api적용시 삭제
+
 
 const MainPage = ({ navigation }) => {
-	//const [books, setBooks] = useState([]);
-	//const [notes, setNotes] = useState([]);
-	//const handleContents = (books, notes) => {
-		//setBooks(books);
-		//setNotes(notes);
-	//};
+	const [books, setBooks] = useState([]);
+	const [notes, setNotes] = useState([]);
+	const handleContents = (newBooks, newNotes) => {
+		setBooks(newBooks);
+		setNotes(newNotes);
+	};
+	useEffect(() => {
+		getMain(handleContents);
+	}, []);
 	const [pnum1, setPnum1] = useState(0);
 	const [pnum2, setPnum2] = useState(5);
 	const [pnum3, setPnum3] = useState(0);
@@ -137,9 +126,7 @@ const MainPage = ({ navigation }) => {
 	HandleSearch = (text) => {
 		setSearchContent("검색: " + text);
 	};
-	//useEffect(() => {
-		//getMain(handleContents);
-	//}, []);
+
 	return (
 		<Container>
 			<CommunityHeader navigation={navigation} />
